Migrate shared and engine support modules to TypeScript
