+++ tsconfig.json
@@ -1,0 +1,58 @@
+{
+  "compilerOptions": {
+    "target": "ES2022",
+    "module": "Node16",
+    "moduleResolution": "Node16",
+    "strict": true,
+    "allowJs": true,
+    "checkJs": false,
+    "declaration": true,
+    "incremental": true,
+    "outDir": ".tsbuild",
+    "rootDir": ".",
+    "lib": ["ES2022"],
+    "types": ["node"],
+    "skipLibCheck": true
+  },
+  "include": [
+    "backend/**/*",
+    "shared/**/*",
+    "scripts/**/*",
+    "tests/gameplay/**/*"
+  ],
+  "exclude": [
+    "node_modules",
+    "frontend",
+    "e2e",
+    ".git",
+    ".tsbuild",
+    "playwright-report",
+    "test-results",
+    "data",
+    "shared/core-domain.cjs",
+    "shared/game-actions.cjs",
+    "shared/messages.cjs",
+    "shared/dice.cjs",
+    "shared/cards.cjs",
+    "shared/models.cjs",
+    "shared/map-graph.cjs",
+    "shared/map-loader.cjs",
+    "shared/continent-loader.cjs",
+    "shared/maps/index.cjs",
+    "shared/maps/classic-mini.cjs",
+    "shared/maps/middle-earth.cjs",
+    "shared/maps/world-classic.cjs",
+    "backend/engine/combat-dice.cjs",
+    "backend/engine/game-setup.cjs",
+    "backend/engine/reinforcement-calculator.cjs",
+    "backend/engine/reinforcement-placement.cjs",
+    "backend/engine/combat-resolution.cjs",
+    "backend/engine/attack-validation.cjs",
+    "backend/engine/conquest-resolution.cjs",
+    "backend/engine/banzai-attack.cjs",
+    "backend/engine/victory-detection.cjs",
+    "backend/engine/fortify-movement.cjs",
+    "backend/engine/ai-player.cjs",
+    "backend/new-game-config.cjs"
+  ]
+}
